DRR - Cpptraj: Move sander energy class to separate file since it really doesnt have much in common with the existing energy class.
