+++ src/Energy.h
@@ -2,9 +2,6 @@
 #define INC_ENERGY_H
 #include "Topology.h"
 #include "Timer.h"
-#ifdef USE_SANDERLIB
-#  include "sander.h"
-#endif
 /// Calculate energy/force from coordinates.
 class Energy_Amber {
   public:
@@ -37,26 +34,4 @@ class Energy_Amber {
     Timer time_14_;
     Timer time_NB_;
 };
-#ifdef USE_SANDERLIB
-class Energy_Sander {
-  public:
-    Energy_Sander() : top_pindex_(-1) {}
-    ~Energy_Sander();
-    int Initialize(Topology const&, Frame&);
-    int CalcEnergy(Topology const&, Frame&);
-    double Ebond() const { return energy_.bond; }
-    double Eangle() const { return energy_.angle; }
-    double Edihedral() const { return energy_.dihedral; }
-    double Evdw14() const { return energy_.vdw_14; }
-    double Eelec14() const { return energy_.elec_14; }
-    double Evdw() const { return energy_.vdw; }
-    double Eelec() const { return energy_.elec; } 
-  private:
-    sander_input input_;
-    pot_ene energy_;
-    FileName top_filename_;
-    std::vector<double> forces_;
-    int top_pindex_;
-};
-#endif /* USE_SANDERLIB */
 #endif
