+++ src/cpptrajfiles
@@ -199,6 +199,7 @@ SOURCES=ActionFrameCounter.cpp \
         EnsembleOut_Multi.cpp \
         EnsembleOut_Single.cpp \
         EnsembleOutList.cpp \
+        Esander.cpp \
         Exec_Calc.cpp \
         Exec_CombineCoords.cpp \
         Exec_Commands.cpp \
